Portfolio Content Page Update
With the typography and image/content behavior at an acceptable level,
the project is ready for the inclusion of additional portfolio content
pages.

Image loading still needs to be made more efficient--choosing smaller
size photos for mobile devices, for instance. The "srcset" artifact from
a previous project can be seen on this current version of "s4p.html".

There are still no visual effects on the loading of the webpage. Those
will come later.