Substantial index.html update and nav update
Adding a descriptive overlay to each portfolio icon on the index page,
as well as cleaning up the navigation to make it more mobile-friendly.

jQuery solutions need to be engineered to control animation of overlay
elements on both mobile and desktop. For now, overlays will remain
active on mobile and will appear on hover for desktop.

Navigation substantially improved for mobile. More work needs to be done
on all screen modes to improve usability.